Feed editing

+++ Primal/State/Model/PostRequestResult.swift
@@ -27,5 +27,5 @@ struct NostrRepost {
 
 public struct PopularHashtag {
     var title: String
-    var apperances: Int
+    var apperances: Double
 }
